requirements/base.txt
Django==2.2.3
django-environ==0.4.5
django-tastypie==0.9.15
Pillow==6.0.0
psycopg2==2.8.3
setproctitle==1.1.10
django-prometheus==1.0.15
